+++ packages/gitbook/src/components/AIActions/AIActionsDropdown.tsx
@@ -6,8 +6,9 @@ import {
     OpenInLLM,
     ViewAsMarkdown,
 } from '@/components/AIActions/AIActions';
-import { Button } from '@/components/primitives/Button';
-import { DropdownMenu } from '@/components/primitives/DropdownMenu';
+import { Button, ButtonGroup } from '@/components/primitives/Button';
+import { DropdownMenu, DropdownMenuSeparator } from '@/components/primitives/DropdownMenu';
+import type { SiteCustomizationSettings } from '@gitbook/api';
 
 import { Icon } from '@gitbook/icons';
 import { useRef } from 'react';
@@ -16,10 +17,7 @@ interface AIActionsDropdownProps {
     markdownPageUrl: string;
     withAIChat?: boolean;
     trademark: boolean;
-    /**
-     * Whether to include the "Open in LLM" entries in the dropdown menu.
-     */
-    withLLMActions?: boolean;
+    actions: SiteCustomizationSettings['pageActions'];
 }
 
 /**
@@ -29,29 +27,31 @@ export function AIActionsDropdown(props: AIActionsDropdownProps) {
     const ref = useRef<HTMLDivElement>(null);
 
     return (
-        <div ref={ref} className="flex h-fit items-stretch justify-start">
+        <ButtonGroup ref={ref}>
             <DefaultAction {...props} />
-            <DropdownMenu
-                align="end"
-                className="!min-w-60 max-w-max"
-                button={
-                    <Button
-                        icon={
-                            <Icon
-                                icon="chevron-down"
-                                className="size-3 transition-transform group-data-[state=open]/button:rotate-180"
-                            />
-                        }
-                        iconOnly
-                        size="xsmall"
-                        variant="secondary"
-                        className="hover:!scale-100 !shadow-none !rounded-l-none bg-tint-base text-sm"
-                    />
-                }
-            >
-                <AIActionsDropdownMenuContent {...props} />
-            </DropdownMenu>
-        </div>
+            {props.actions.markdown || props.actions.externalAI ? (
+                <DropdownMenu
+                    align="end"
+                    className="!min-w-60 max-w-max"
+                    button={
+                        <Button
+                            icon={
+                                <Icon
+                                    icon="chevron-down"
+                                    className="size-3 transition-transform group-data-[state=open]/button:rotate-180"
+                                />
+                            }
+                            iconOnly
+                            size="xsmall"
+                            variant="secondary"
+                            className="bg-tint-base text-sm"
+                        />
+                    }
+                >
+                    <AIActionsDropdownMenuContent {...props} />
+                </DropdownMenu>
+            ) : null}
+        </ButtonGroup>
     );
 }
 
@@ -59,7 +59,7 @@ export function AIActionsDropdown(props: AIActionsDropdownProps) {
  * The content of the dropdown menu.
  */
 function AIActionsDropdownMenuContent(props: AIActionsDropdownProps) {
-    const { markdownPageUrl, withAIChat, trademark, withLLMActions } = props;
+    const { markdownPageUrl, withAIChat, trademark, actions } = props;
 
     return (
         <>
@@ -67,15 +67,21 @@ function AIActionsDropdownMenuContent(props: AIActionsDropdownProps) {
                 <OpenDocsAssistant trademark={trademark} type="dropdown-menu-item" />
             ) : null}
 
-            <CopyMarkdown
-                isDefaultAction={!withAIChat}
-                markdownPageUrl={markdownPageUrl}
-                type="dropdown-menu-item"
-            />
-            <ViewAsMarkdown markdownPageUrl={markdownPageUrl} type="dropdown-menu-item" />
-
-            {withLLMActions ? (
+            {actions.markdown ? (
                 <>
+                    <DropdownMenuSeparator className="first:hidden" />
+                    <CopyMarkdown
+                        isDefaultAction={!withAIChat}
+                        markdownPageUrl={markdownPageUrl}
+                        type="dropdown-menu-item"
+                    />
+                    <ViewAsMarkdown markdownPageUrl={markdownPageUrl} type="dropdown-menu-item" />
+                </>
+            ) : null}
+
+            {actions.externalAI ? (
+                <>
+                    <DropdownMenuSeparator className="first:hidden" />
                     <OpenInLLM provider="chatgpt" url={markdownPageUrl} type="dropdown-menu-item" />
                     <OpenInLLM provider="claude" url={markdownPageUrl} type="dropdown-menu-item" />
                 </>
@@ -88,17 +94,27 @@ function AIActionsDropdownMenuContent(props: AIActionsDropdownProps) {
  * A default action shown as a quick-access button beside the dropdown menu
  */
 function DefaultAction(props: AIActionsDropdownProps) {
-    const { markdownPageUrl, withAIChat, trademark } = props;
+    const { markdownPageUrl, withAIChat, trademark, actions } = props;
 
     if (withAIChat) {
         return <OpenDocsAssistant trademark={trademark} type="button" />;
     }
 
-    return (
-        <CopyMarkdown
-            isDefaultAction={!withAIChat}
-            markdownPageUrl={markdownPageUrl}
-            type="button"
-        />
-    );
+    if (actions.markdown) {
+        return (
+            <CopyMarkdown
+                isDefaultAction={!withAIChat}
+                markdownPageUrl={markdownPageUrl}
+                type="button"
+            />
+        );
+    }
+
+    if (actions.externalAI) {
+        return (
+            <>
+                <OpenInLLM provider="chatgpt" url={markdownPageUrl} type="button" />
+            </>
+        );
+    }
 }