+++ packages/gitbook/src/components/DocumentView/Heading.tsx
@@ -24,28 +24,32 @@ export function Heading(props: BlockProps<DocumentBlockHeading>) {
             className={tcls(
                 textStyle.textSize,
                 'heading',
-                'grid',
-                'scroll-m-12',
+                'flex',
+                'items-baseline',
+                'scroll-mt-(--content-scroll-margin)',
+                getTextAlignment(block.data.align),
                 hashLinkButtonWrapperStyles,
-                style
+                style,
+                textStyle.marginTop
             )}
         >
             <HashLinkButton
                 id={id}
                 block={block}
-                className={tcls('-ml-6', textStyle.anchorButtonMarginTop)}
+                className={tcls('-ml-6', 'pr-2')}
                 iconClassName={tcls('size-4')}
                 label="Direct link to heading"
             />
 
             <div
                 className={tcls(
-                    'grid-area-1-1',
-                    'z-[1]',
+                    'flex-1',
+                    'z-1',
                     'justify-self-start',
+                    'max-w-full',
+                    'break-words',
                     getTextAlignment(block.data.align),
-                    textStyle.lineHeight,
-                    textStyle.marginTop
+                    textStyle.lineHeight
                 )}
             >
                 <Inlines {...rest} context={context} nodes={block.nodes} ancestorInlines={[]} />