+++ packages/gitbook/src/components/DocumentView/Text.tsx
@@ -71,11 +71,7 @@ function Strikethrough(props: MarkedLeafProps<DocumentMarkStrikethrough>) {
 }
 
 function Keyboard(props: MarkedLeafProps<DocumentMarkKeyboard>) {
-    return (
-        <kbd className="rounded border px-1 font-mono shadow-[0_1px_0_0_theme(borderColor.DEFAULT)]">
-            {props.children}
-        </kbd>
-    );
+    return <kbd className="rounded-xs border border-b-2 px-1 font-mono">{props.children}</kbd>;
 }
 
 function Superscript(props: MarkedLeafProps<DocumentMarkSuperscript>) {
@@ -90,9 +86,9 @@ function Code(props: MarkedLeafProps<DocumentMarkCode>) {
     return (
         <code
             className={tcls(
-                'py-[1px]',
+                'py-px',
                 'px-1.5',
-                'min-w-[1.625rem]',
+                'min-w-6.5',
                 'justify-center',
                 'items-center',
                 'leading-normal',
@@ -100,7 +96,7 @@ function Code(props: MarkedLeafProps<DocumentMarkCode>) {
                 'ring-inset',
                 'ring-tint',
                 'bg-tint',
-                'rounded',
+                'rounded-sm',
                 // Text size is proportional to the font-size of the parent element
                 'text-[.875em]',
                 // We ensure that the code is not making the parent bigger, especially in headings
@@ -126,7 +122,25 @@ function Color(props: MarkedLeafProps<DocumentMarkColor>) {
     );
 }
 
-const textColorToStyle: { [color in DocumentMarkColor['data']['text']]: ClassValue } = {
+/**
+ * @TODO replace by DocumentMarkColor['data']['text'] and DocumentMarkColor['data']['background']
+ * once the API is updated.
+ */
+type DocumentMarkColorValue =
+    | 'default'
+    | 'green'
+    | 'blue'
+    | 'red'
+    | 'orange'
+    | 'yellow'
+    | 'purple'
+    | '$primary'
+    | '$info'
+    | '$success'
+    | '$warning'
+    | '$danger';
+
+const textColorToStyle: { [color in DocumentMarkColorValue]: ClassValue } = {
     default: [],
     blue: ['text-blue-500'],
     red: ['text-red-500'],
@@ -134,9 +148,14 @@ const textColorToStyle: { [color in DocumentMarkColor['data']['text']]: ClassVal
     yellow: ['text-yellow-600'],
     purple: ['text-purple-500'],
     orange: ['text-orange-500'],
+    $primary: ['text-primary'],
+    $info: ['text-info'],
+    $success: ['text-success'],
+    $warning: ['text-warning'],
+    $danger: ['text-danger'],
 };
 
-const backgroundColorToStyle: { [color in DocumentMarkColor['data']['background']]: ClassValue } = {
+const backgroundColorToStyle: { [color in DocumentMarkColorValue]: ClassValue } = {
     default: [],
     blue: ['bg-mark-blue'],
     red: ['bg-mark-red'],
@@ -144,4 +163,9 @@ const backgroundColorToStyle: { [color in DocumentMarkColor['data']['background'
     yellow: ['bg-mark-yellow'],
     purple: ['bg-mark-purple'],
     orange: ['bg-mark-orange'],
+    $primary: ['bg-primary'],
+    $info: ['bg-info'],
+    $success: ['bg-success'],
+    $warning: ['bg-warning'],
+    $danger: ['bg-danger'],
 };