+++ packages/gitbook/src/components/Search/SearchAskAnswer.tsx
@@ -10,11 +10,12 @@ import { t } from '@/intl/translate';
 import type { TranslationLanguage } from '@/intl/translations';
 import { tcls } from '@/lib/tailwind';
 
+import { AIResponseFeedback } from '../AIChat';
 import { useTrackEvent } from '../Insights';
 import { Link } from '../primitives';
 import { useSearchAskContext } from './SearchAskContext';
 import { type AskAnswerResult, type AskAnswerSource, streamAskQuestion } from './server-actions';
-import { useSearch, useSearchLink } from './useSearch';
+import { useSearchLink } from './useSearch';
 
 export type SearchAskState =
     | {
@@ -36,7 +37,6 @@ export function SearchAskAnswer(props: { query: string }) {
 
     const language = useLanguage();
     const trackEvent = useTrackEvent();
-    const [, setSearchState] = useSearch();
     const [askState, setAskState] = useSearchAskContext();
 
     React.useEffect(() => {
@@ -49,10 +49,6 @@ export function SearchAskAnswer(props: { query: string }) {
                 type: 'ask_question',
                 query,
             });
-
-            // When we pass in "ask" mode, the query could still be updated by the client
-            // we ensure that the query is up-to-date before starting the stream.
-            setSearchState((prev) => (prev ? { ...prev, query, ask: true } : null));
 
             const { stream } = await streamAskQuestion({ question: query });
             for await (const chunk of readStreamableValue(stream)) {
@@ -79,7 +75,7 @@ export function SearchAskAnswer(props: { query: string }) {
                 cancelled = true;
             }
         };
-    }, [query, setAskState, setSearchState, trackEvent]);
+    }, [query, setAskState, trackEvent]);
 
     React.useEffect(() => {
         return () => {
@@ -97,7 +93,11 @@ export function SearchAskAnswer(props: { query: string }) {
         <div className="flex min-h-full p-4">
             {askState?.type === 'answer' ? (
                 <React.Suspense fallback={loading}>
-                    <TransitionAnswerBody answer={askState.answer} placeholder={loading} />
+                    <TransitionAnswerBody
+                        answer={askState.answer}
+                        placeholder={loading}
+                        query={query}
+                    />
                 </React.Suspense>
             ) : null}
             {askState?.type === 'error' ? (
@@ -114,8 +114,12 @@ export function SearchAskAnswer(props: { query: string }) {
  * Since the answer can be an async component that could suspend rendering,
  * we need to wrap it in a transition to avoid flickering.
  */
-function TransitionAnswerBody(props: { answer: AskAnswerResult; placeholder: React.ReactNode }) {
-    const { answer, placeholder } = props;
+function TransitionAnswerBody(props: {
+    query: string;
+    answer: AskAnswerResult;
+    placeholder: React.ReactNode;
+}) {
+    const { query, answer, placeholder } = props;
     const [display, setDisplay] = React.useState<AskAnswerResult | null>(null);
     const [_isPending, startTransition] = React.useTransition();
 
@@ -127,21 +131,25 @@ function TransitionAnswerBody(props: { answer: AskAnswerResult; placeholder: Rea
 
     return display ? (
         <div className={tcls('w-full')}>
-            <AnswerBody answer={display} />
+            <AnswerBody query={query} answer={display} />
         </div>
     ) : (
         <>{placeholder}</>
     );
 }
 
-function AnswerBody(props: { answer: AskAnswerResult }) {
-    const { answer } = props;
+function AnswerBody(props: { query: string; answer: AskAnswerResult }) {
+    const { query, answer } = props;
     const language = useLanguage();
 
     return (
         <>
             <div data-testid="search-ask-answer" className="text-tint-strong">
                 {answer.body ?? t(language, 'search_ask_no_answer')}
+                {answer.sources.length > 0 ? (
+                    // @TODO: Add responseId once search uses new AI endpoint
+                    <AIResponseFeedback query={query} className="-ml-1 mt-2" responseId="" />
+                ) : null}
                 {answer.followupQuestions.length > 0 ? (
                     <AnswerFollowupQuestions followupQuestions={answer.followupQuestions} />
                 ) : null}
@@ -162,7 +170,7 @@ function AnswerFollowupQuestions(props: { followupQuestions: string[] }) {
     const getSearchLinkProps = useSearchLink();
 
     return (
-        <div className={tcls('flex', 'flex-col', 'flex-wrap', 'mt-4', 'sm:mt-6')}>
+        <div className={tcls('flex', 'flex-col', 'flex-wrap', 'mt-4')}>
             {followupQuestions.map((question) => (
                 <Link
                     key={question}
@@ -173,7 +181,7 @@ function AnswerFollowupQuestions(props: { followupQuestions: string[] }) {
                         'px-4',
                         '-mx-4',
                         'py-2',
-                        'rounded',
+                        'rounded-sm',
                         'straight-corners:rounded-none',
                         'circular-corners:rounded-full',
                         'text-tint',
@@ -181,8 +189,7 @@ function AnswerFollowupQuestions(props: { followupQuestions: string[] }) {
                         'focus-within:bg-tint-hover'
                     )}
                     {...getSearchLinkProps({
-                        query: question,
-                        ask: true,
+                        ask: question,
                     })}
                 >
                     <Icon