+++ packages/gitbook/src/components/primitives/Button.tsx
@@ -30,52 +30,57 @@ export const variantClasses = {
         'hover:bg-primary-solid-hover',
         'hover:text-contrast-primary-solid-hover',
         'border-0',
-        'contrast-more:border-1',
+        'contrast-more:border',
+        'disabled:bg-tint',
+        'disabled:text-tint/8',
     ],
     blank: [
         'bg-transparent',
         'text-tint',
         'border-0',
-        'shadow-none',
+        'shadow-none!',
         'hover:bg-primary-hover',
         'hover:text-primary-strong',
-        'hover:scale-1',
         'contrast-more:bg-tint-subtle',
-        'depth-subtle:hover:translate-y-0',
+        'hover:depth-subtle:translate-y-0',
+        'disabled:text-tint/8',
+        'disabled:bg-transparent',
     ],
     secondary: [
         'bg-tint',
         'depth-flat:bg-transparent',
         'text-tint',
         'hover:bg-tint-hover',
-        'depth-flat:hover:bg-tint-hover',
+        'hover:depth-flat:bg-tint-hover',
         'hover:text-primary',
         'contrast-more:bg-tint-subtle',
+        'disabled:bg-transparent',
+        'disabled:text-tint/8',
     ],
     header: [
         'bg-tint-base text-tint',
-        'theme-clean:hover:bg-tint-subtle',
+        'hover:theme-clean:bg-tint-subtle',
 
         'theme-bold:bg-header-link/2',
         'theme-bold:text-header-link',
-        'theme-bold:!shadow-none',
-        'theme-bold:ring-header-link/4',
+        'theme-bold:shadow-none!',
+        'theme-bold:border-header-link/4',
 
-        'theme-bold:hover:bg-header-link/3',
-        'theme-bold:hover:text-header-link',
-        'theme-bold:hover:shadow-none',
-        'theme-bold:hover:ring-header-link/5',
+        'hover:theme-bold:bg-header-link/3',
+        'hover:theme-bold:text-header-link',
+        'hover:theme-bold:shadow-none',
+        'hover:theme-bold:border-header-link/5',
 
-        'theme-bold:contrast-more:bg-header-background',
-        'theme-bold:contrast-more:text-header-link',
-        'theme-bold:contrast-more:ring-header-link',
-        'theme-bold:contrast-more:hover:ring-header-link',
+        'contrast-more:theme-bold:bg-header-background',
+        'contrast-more:theme-bold:text-header-link',
+        'contrast-more:theme-bold:border-header-link',
+        'contrast-more:hover:theme-bold:border-header-link',
     ],
 };
 
 const activeClasses = {
     primary: 'bg-primary-solid-hover',
-    blank: 'bg-primary-active text-primary-strong font-medium hover:text-primary-strong hover:bg-primary-active',
+    blank: 'bg-primary-active disabled:bg-primary-active text-primary-strong font-medium hover:text-primary-strong disabled:text-primary-strong hover:bg-primary-active',
     secondary: 'bg-tint-active',
     header: 'bg-header-link/3',
 };
@@ -99,13 +104,14 @@ export const Button = React.forwardRef<
             children,
             active,
             trailing,
+            disabled,
             ...rest
         },
         ref
     ) => {
         const sizes = {
-            default: ['text-base', 'font-semibold', 'px-5', 'py-2', 'circular-corners:px-6'],
-            medium: ['text-sm', 'px-3.5', 'py-1.5', 'circular-corners:px-4'],
+            default: ['text-base', 'font-semibold', iconOnly ? 'px-2' : 'px-5', 'py-2'],
+            medium: ['text-sm', iconOnly ? 'px-2' : 'px-3.5', 'py-1.5'],
             small: ['text-xs', 'py-2', iconOnly ? 'px-2' : 'px-3'],
             xsmall: ['text-xs', 'py-1', iconOnly ? 'px-1.5' : 'px-2'],
         };
@@ -142,6 +148,7 @@ export const Button = React.forwardRef<
                     classNames={['ButtonStyles']}
                     insights={insights}
                     aria-label={label?.toString()}
+                    aria-pressed={active === undefined ? undefined : active}
                     target={target}
                     {...rest}
                 >
@@ -156,12 +163,38 @@ export const Button = React.forwardRef<
                 type="button"
                 className={tcls(buttonOnlyClassNames, domClassName)}
                 aria-label={label?.toString()}
+                aria-pressed={active === undefined ? undefined : active}
+                disabled={disabled}
                 {...rest}
             >
                 {content}
             </button>
         );
 
-        return iconOnly && label ? <Tooltip label={label}>{button}</Tooltip> : button;
+        return iconOnly && label ? (
+            <Tooltip
+                rootProps={{ open: disabled === true ? false : undefined }}
+                label={label}
+                triggerProps={{ disabled }}
+            >
+                {button}
+            </Tooltip>
+        ) : (
+            button
+        );
     }
 );
+
+export const ButtonGroup = React.forwardRef<HTMLDivElement, ButtonProps>(({ children }, ref) => {
+    return (
+        <div
+            ref={ref}
+            className={tcls(
+                'flex h-fit items-stretch justify-start *:translate-y-0! *:shadow-none!',
+                '[&>*:not(:first-child)]:border-l-0 [&>*:not(:first-child,:last-child)]:rounded-none [&>*:not(:only-child):first-child]:rounded-r-none [&>*:not(:only-child):last-child]:rounded-l-none'
+            )}
+        >
+            {children}
+        </div>
+    );
+});