+++ packages/openapi-parser/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @gitbook/openapi-parser
 
+## 3.0.0
+
+### Major Changes
+
+- f1a6dec: Less strict OpenAPI parser, now returns { errors, filesystem }
+
+## 2.2.2
+
+### Patch Changes
+
+- b5ad0ce: Fix OpenAPI v2 parsing
+
+## 2.2.1
+
+### Patch Changes
+
+- 1738677: Stricter validation
+
 ## 2.2.0
 
 ### Minor Changes