+++ packages/react-contentkit/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @gitbook/react-contentkit
 
+## 0.7.3
+
+### Patch Changes
+
+- Updated dependencies [2cdba53]
+  - @gitbook/icons@0.2.2
+
+## 0.7.2
+
+### Patch Changes
+
+- Updated dependencies [6816f0f]
+  - @gitbook/icons@0.2.1
+
 ## 0.7.1
 
 ### Patch Changes